refactor: moving all ES6 files to src/, working on #83
