+++ .npmignore
@@ -1,1 +1,4 @@
-test/
+/coverage
+/src
+/test
+.*
