+++ src/tags/capture.js
@@ -2,8 +2,8 @@
 const Liquid = require('..')
 const lexical = Liquid.lexical
 const re = new RegExp(`(${lexical.identifier.source})`)
-const assert = require('../src/util/assert.js')
-const types = require('../src/scope.js').types
+const assert = require('../util/assert.js')
+const types = require('../scope.js').types
 
 module.exports = function (liquid) {
   liquid.registerTag('capture', {
