+++ src/tags/include.js
@@ -3,7 +3,7 @@ const Liquid = require('..')
 const lexical = Liquid.lexical
 const withRE = new RegExp(`with\\s+(${lexical.value.source})`)
 const staticFileRE = /[^\s,]+/
-const assert = require('../src/util/assert.js')
+const assert = require('../util/assert.js')
 
 module.exports = function (liquid) {
   liquid.registerTag('include', {
